Update: Split out generic AST methods into utility (fixes #962)
