+++ lib/eslint.js
@@ -20,7 +20,8 @@ var estraverse = require("estraverse-fb"),
     EventEmitter = require("events").EventEmitter,
     escapeRegExp = require("escape-string-regexp"),
     validator = require("./config-validator"),
-    replacements = require("../conf/replacements.json");
+    replacements = require("../conf/replacements.json"),
+    astUtils = require("./ast-utils");
 
 //------------------------------------------------------------------------------
 // Helpers
@@ -899,27 +900,7 @@ module.exports = (function() {
      * @param {ASTNode} node The AST node to get the comments for.
      * @returns {Object} The list of comments indexed by their position.
      */
-    api.getComments = function(node) {
-
-        var leadingComments = node.leadingComments || [],
-            trailingComments = node.trailingComments || [];
-
-        /*
-         * espree adds a "comments" array on Program nodes rather than
-         * leadingComments/trailingComments. Comments are only left in the
-         * Program node comments array if there is no executable code.
-         */
-        if (node.type === "Program") {
-            if (node.body.length === 0) {
-                leadingComments = node.comments;
-            }
-        }
-
-        return {
-            leading: leadingComments,
-            trailing: trailingComments
-        };
-    };
+    api.getComments = astUtils.getComments;
 
     /**
      * Retrieves the JSDoc comment for a given node.
@@ -927,71 +908,7 @@ module.exports = (function() {
      * @returns {ASTNode} The BlockComment node containing the JSDoc for the
      *      given node or null if not found.
      */
-    api.getJSDocComment = function(node) {
-
-        var parent = node.parent,
-            line = node.loc.start.line;
-
-        /**
-         * Finds a JSDoc comment node in an array of comment nodes.
-         * @param {ASTNode[]} comments The array of comment nodes to search.
-         * @returns {ASTNode} The node if found, null if not.
-         * @private
-         */
-        function findJSDocComment(comments) {
-
-            if (comments) {
-                for (var i = comments.length - 1; i >= 0; i--) {
-                    if (comments[i].type === "Block" && comments[i].value.charAt(0) === "*") {
-
-                        if (line - comments[i].loc.end.line <= 1) {
-                            return comments[i];
-                        } else {
-                            break;
-                        }
-                    }
-                }
-            }
-
-            return null;
-        }
-
-        /**
-         * Check to see if its a ES6 export declaration
-         * @param {ASTNode} astNode - any node
-         * @returns {boolean} whether the given node represents a export declaration
-         */
-        function looksLikeExport(astNode) {
-            return astNode.type === "ExportDefaultDeclaration" || astNode.type === "ExportNamedDeclaration" ||
-                astNode.type === "ExportAllDeclaration" || astNode.type === "ExportSpecifier";
-        }
-
-        switch (node.type) {
-            case "FunctionDeclaration":
-                if (looksLikeExport(parent)) {
-                    return findJSDocComment(parent.leadingComments);
-                } else {
-                    return findJSDocComment(node.leadingComments);
-                }
-                break;
-
-            case "ArrowFunctionExpression":
-            case "FunctionExpression":
-
-                if (parent.type !== "CallExpression") {
-                    while (parent && !parent.leadingComments && !/Function/.test(parent.type)) {
-                        parent = parent.parent;
-                    }
-
-                    return parent && (parent.type !== "FunctionDeclaration") ? findJSDocComment(parent.leadingComments) : null;
-                }
-
-                // falls through
-
-            default:
-                return null;
-        }
-    };
+    api.getJSDocComment = astUtils.getJSDocComment;
 
     /**
      * Gets nodes that are ancestors of current node.
