+++ lib/rules/array-bracket-spacing.js
@@ -8,6 +8,8 @@
  */
 "use strict";
 
+var astUtils = require("../ast-utils");
+
 //------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
@@ -36,26 +38,6 @@ module.exports = function(context) {
     //--------------------------------------------------------------------------
     // Helpers
     //--------------------------------------------------------------------------
-
-    /**
-     * Determines whether two adjacent tokens are have whitespace between them.
-     * @param {Object} left - The left token object.
-     * @param {Object} right - The right token object.
-     * @returns {boolean} Whether or not there is space between the tokens.
-     */
-    function isSpaced(left, right) {
-        return left.range[1] < right.range[0];
-    }
-
-    /**
-     * Determines whether two adjacent tokens are on the same line.
-     * @param {Object} left - The left token object.
-     * @param {Object} right - The right token object.
-     * @returns {boolean} Whether or not the tokens are on the same line.
-     */
-    function isSameLine(left, right) {
-        return left.loc.start.line === right.loc.start.line;
-    }
 
     /**
     * Reports that there shouldn't be a space after the first token
@@ -128,20 +110,20 @@ module.exports = function(context) {
             options.singleElementException && node.elements.length === 1
                 ? !options.spaced : options.spaced;
 
-        if (isSameLine(first, second)) {
-            if (openingBracketMustBeSpaced && !isSpaced(first, second)) {
+        if (astUtils.isTokenOnSameLine(first, second)) {
+            if (openingBracketMustBeSpaced && !astUtils.isTokenSpaced(first, second)) {
                 reportRequiredBeginningSpace(node, first);
             }
-            if (!openingBracketMustBeSpaced && isSpaced(first, second)) {
+            if (!openingBracketMustBeSpaced && astUtils.isTokenSpaced(first, second)) {
                 reportNoBeginningSpace(node, first);
             }
         }
 
-        if (isSameLine(penultimate, last)) {
-            if (closingBracketMustBeSpaced && !isSpaced(penultimate, last)) {
+        if (astUtils.isTokenOnSameLine(penultimate, last)) {
+            if (closingBracketMustBeSpaced && !astUtils.isTokenSpaced(penultimate, last)) {
                 reportRequiredEndingSpace(node, last);
             }
-            if (!closingBracketMustBeSpaced && isSpaced(penultimate, last)) {
+            if (!closingBracketMustBeSpaced && astUtils.isTokenSpaced(penultimate, last)) {
                 reportNoEndingSpace(node, last);
             }
         }
