+++ lib/rules/comma-spacing.js
@@ -5,6 +5,8 @@
  */
 "use strict";
 
+var astUtils = require("../ast-utils");
+
 //------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
@@ -52,17 +54,6 @@ module.exports = function(context) {
     }
 
     /**
-     * Checks whether two tokens are on the same line.
-     * @param {ASTNode} left The leftmost token.
-     * @param {ASTNode} right The rightmost token.
-     * @returns {boolean} True if the tokens are on the same line, false if not.
-     * @private
-     */
-    function isSameLine(left, right) {
-        return left.loc.end.line === right.loc.start.line;
-    }
-
-    /**
      * Determines if a given token is a comma operator.
      * @param {ASTNode} token The token to check.
      * @returns {boolean} True if the token is a comma, false if not.
@@ -96,12 +87,12 @@ module.exports = function(context) {
      * @private
      */
     function validateCommaItemSpacing(tokens, reportItem) {
-        if (tokens.left && isSameLine(tokens.left, tokens.comma) &&
+        if (tokens.left && astUtils.isTokenOnSameLine(tokens.left, tokens.comma) &&
                 (options.before !== isSpaced(tokens.left, tokens.comma))
         ) {
             report(reportItem, "before");
         }
-        if (tokens.right && isSameLine(tokens.comma, tokens.right) &&
+        if (tokens.right && astUtils.isTokenOnSameLine(tokens.comma, tokens.right) &&
                 (options.after !== isSpaced(tokens.comma, tokens.right))
         ) {
             report(reportItem, "after");
