+++ lib/rules/no-ex-assign.js
@@ -5,6 +5,8 @@
 
 "use strict";
 
+var astUtils = require("../ast-utils");
+
 //------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
@@ -12,35 +14,16 @@
 module.exports = function(context) {
 
     /**
-     * Reports a reference if is non initializer and writable.
-     * @param {Reference} reference - A reference to check.
-     * @param {int} index - The index of the reference in the references.
-     * @param {Reference[]} references - The array that the reference belongs to.
-     * @returns {void}
-     */
-    function checkReference(reference, index, references) {
-        var identifier = reference.identifier;
-
-        if (identifier != null &&
-            reference.init === false &&
-            reference.isWrite() &&
-            // Destructuring assignments can have multiple default value,
-            // so possibly there are multiple writeable references for the same identifier.
-            (index === 0 || references[index - 1].identifier !== identifier)
-        ) {
-            context.report(
-                identifier,
-                "Do not assign to the exception parameter.");
-        }
-    }
-
-    /**
      * Finds and reports references that are non initializer and writable.
      * @param {Variable} variable - A variable to check.
      * @returns {void}
      */
     function checkVariable(variable) {
-        variable.references.forEach(checkReference);
+        astUtils.getModifyingReferences(variable.references).forEach(function(reference) {
+            context.report(
+                reference.identifier,
+                "Do not assign to the exception parameter.");
+        });
     }
 
     return {
