+++ lib/rules/no-func-assign.js
@@ -6,6 +6,8 @@
 
 "use strict";
 
+var astUtils = require("../ast-utils");
+
 //------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
@@ -36,26 +38,16 @@ module.exports = function(context) {
 
     /**
      * Reports a reference if is non initializer and writable.
-     * @param {Reference} reference - A reference to check.
-     * @param {int} index - The index of the reference in the references.
-     * @param {Reference[]} references - The array that the reference belongs to.
+     * @param {References} references - Collection of reference to check.
      * @returns {void}
      */
-    function checkReference(reference, index, references) {
-        var identifier = reference.identifier;
-
-        if (identifier != null &&
-            reference.init === false &&
-            reference.isWrite() &&
-            // Destructuring assignments can have multiple default value,
-            // so possibly there are multiple writeable references for the same identifier.
-            (index === 0 || references[index - 1].identifier !== identifier)
-        ) {
+    function checkReference(references) {
+        astUtils.getModifyingReferences(references).forEach(function(reference) {
             context.report(
-                identifier,
+                reference.identifier,
                 "'{{name}}' is a function.",
-                {name: identifier.name});
-        }
+                {name: reference.identifier.name});
+        });
     }
 
     /**
@@ -68,9 +60,9 @@ module.exports = function(context) {
             // If the function is in global scope, its references are not resolved (by escope's design).
             // So when references of the function are nothing, this checks in unresolved.
             if (variable.references.length > 0) {
-                variable.references.forEach(checkReference);
+                checkReference(variable.references);
             } else if (unresolved[variable.name] != null) {
-                unresolved[variable.name].forEach(checkReference);
+                checkReference(unresolved[variable.name]);
             }
         }
     }
