+++ lib/rules/no-invalid-this.js
@@ -15,6 +15,7 @@ var anyFunctionPattern = /^(?:Function(?:Declaration|Expression)|ArrowFunctionEx
 var bindOrCallOrApplyPattern = /^(?:bind|call|apply)$/;
 var arrayOrTypedArrayPattern = /Array$/;
 var arrayMethodPattern = /^(?:every|filter|find|findIndex|forEach|map|some)$/;
+var astUtils = require("../ast-utils");
 
 /**
  * Checks whether or not a node is a constructor.
@@ -31,11 +32,10 @@ function isES5Constructor(node) {
 /**
  * Checks whether or not a node has a `@this` tag in its comments.
  * @param {ASTNode} node - A node to check.
- * @param {RuleContext} context - A context to get the comments of the node.
  * @returns {boolean} Whether or not the node has a `@this` tag in its comments.
  */
-function hasJSDocThisTag(node, context) {
-    return context.getComments(node).leading.some(function(comment) {
+function hasJSDocThisTag(node) {
+    return astUtils.getComments(node).leading.some(function(comment) {
         return thisTagPattern.test(comment.value);
     });
 }
@@ -62,19 +62,6 @@ function getUpperFunction(node) {
  */
 function isCallee(node) {
     return node.parent.type === "CallExpression" && node.parent.callee === node;
-}
-
-/**
- * Checks whether or not a node is `null` or `undefined`.
- * @param {ASTNode} node - A node to check.
- * @returns {boolean} Whether or not the node is a `null` or `undefined`.
- */
-function isNullOrUndefined(node) {
-    return (
-        (node.type === "Literal" && node.value === null) ||
-        (node.type === "Identifier" && node.name === "undefined") ||
-        (node.type === "UnaryExpression" && node.operator === "void")
-    );
 }
 
 /**
@@ -148,11 +135,10 @@ function isMethodWhichHasThisArg(node) {
  * - The function is a callback of array methods (such as `.forEach()`) if `thisArg` is given.
  *
  * @param {ASTNode} node - A node to check.
- * @param {RuleContext} context - A context to get the JSDoc comment of the node.
  * @returns {boolean} A found function node.
  */
-function hasValidThis(node, context) {
-    if (isES5Constructor(node) || hasJSDocThisTag(node, context)) {
+function hasValidThis(node) {
+    if (isES5Constructor(node) || hasJSDocThisTag(node)) {
         return true;
     }
 
@@ -215,7 +201,7 @@ function hasValidThis(node, context) {
                     bindOrCallOrApplyPattern.test(parent.property.name) &&
                     isCallee(parent) &&
                     parent.parent.arguments.length > 0 &&
-                    !isNullOrUndefined(parent.parent.arguments[0])
+                    !astUtils.isNullOrUndefined(parent.parent.arguments[0])
                 );
 
             // e.g.
@@ -227,21 +213,21 @@ function hasValidThis(node, context) {
                     return (
                         parent.arguments.length === 3 &&
                         parent.arguments[0] === node &&
-                        !isNullOrUndefined(parent.arguments[1])
+                        !astUtils.isNullOrUndefined(parent.arguments[1])
                     );
                 }
                 if (isArrayFrom(parent.callee)) {
                     return (
                         parent.arguments.length === 3 &&
                         parent.arguments[1] === node &&
-                        !isNullOrUndefined(parent.arguments[2])
+                        !astUtils.isNullOrUndefined(parent.arguments[2])
                     );
                 }
                 if (isMethodWhichHasThisArg(parent.callee)) {
                     return (
                         parent.arguments.length === 2 &&
                         parent.arguments[0] === node &&
-                        !isNullOrUndefined(parent.arguments[1])
+                        !astUtils.isNullOrUndefined(parent.arguments[1])
                     );
                 }
                 return false;
@@ -276,7 +262,7 @@ module.exports = function(context) {
         var current = this[this.length - 1];
         if (!current.init) {
             current.init = true;
-            current.valid = hasValidThis(current.node, context);
+            current.valid = hasValidThis(current.node);
         }
         return current;
     };
