+++ lib/rules/no-useless-call.js
@@ -6,6 +6,8 @@
 
 "use strict";
 
+var astUtils = require("../ast-utils");
+
 //------------------------------------------------------------------------------
 // Helpers
 //------------------------------------------------------------------------------
@@ -24,19 +26,6 @@ function isCallOrNonVariadicApply(node) {
             (node.callee.property.name === "call" && node.arguments.length >= 1) ||
             (node.callee.property.name === "apply" && node.arguments.length === 2 && node.arguments[1].type === "ArrayExpression")
         )
-    );
-}
-
-/**
- * Checks whether or not a node is `null` or `undefined`.
- * @param {ASTNode} node - A node to check.
- * @returns {boolean} Whether or not the node is a `null` or `undefined`.
- */
-function isNullOrUndefined(node) {
-    return (
-        (node.type === "Literal" && node.value === null) ||
-        (node.type === "Identifier" && node.name === "undefined") ||
-        (node.type === "UnaryExpression" && node.operator === "void")
     );
 }
 
@@ -74,7 +63,7 @@ function equalTokens(left, right, context) {
  */
 function isValidThisArg(expectedThis, thisArg, context) {
     if (expectedThis == null) {
-        return isNullOrUndefined(thisArg);
+        return astUtils.isNullOrUndefined(thisArg);
     }
     return equalTokens(expectedThis, thisArg, context);
 }
