+++ lib/rules/operator-linebreak.js
@@ -6,6 +6,8 @@
 
 "use strict";
 
+var astUtils = require("../ast-utils");
+
 //------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
@@ -17,17 +19,6 @@ module.exports = function(context) {
     //--------------------------------------------------------------------------
     // Helpers
     //--------------------------------------------------------------------------
-
-    /**
-     * Checks whether two tokens are on the same line.
-     * @param {ASTNode} left The leftmost token.
-     * @param {ASTNode} right The rightmost token.
-     * @returns {boolean} True if the tokens are on the same line, false if not.
-     * @private
-     */
-    function isSameLine(left, right) {
-        return left.loc.end.line === right.loc.start.line;
-    }
 
     /**
      * Checks the operator placement
@@ -53,13 +44,13 @@ module.exports = function(context) {
         var operator = operatorToken.value;
 
         // if single line
-        if (isSameLine(leftToken, operatorToken) &&
-                isSameLine(operatorToken, rightToken)) {
+        if (astUtils.isTokenOnSameLine(leftToken, operatorToken) &&
+                astUtils.isTokenOnSameLine(operatorToken, rightToken)) {
 
             return;
 
-        } else if (!isSameLine(leftToken, operatorToken) &&
-                !isSameLine(operatorToken, rightToken)) {
+        } else if (!astUtils.isTokenOnSameLine(leftToken, operatorToken) &&
+                !astUtils.isTokenOnSameLine(operatorToken, rightToken)) {
 
             // lone operator
             context.report(node, {
@@ -67,14 +58,14 @@ module.exports = function(context) {
                 column: operatorToken.loc.end.column
             }, "Bad line breaking before and after '" + operator + "'.");
 
-        } else if (style === "before" && isSameLine(leftToken, operatorToken)) {
+        } else if (style === "before" && astUtils.isTokenOnSameLine(leftToken, operatorToken)) {
 
             context.report(node, {
                 line: operatorToken.loc.end.line,
                 column: operatorToken.loc.end.column
             }, "'" + operator + "' should be placed at the beginning of the line.");
 
-        } else if (style === "after" && isSameLine(operatorToken, rightToken)) {
+        } else if (style === "after" && astUtils.isTokenOnSameLine(operatorToken, rightToken)) {
 
             context.report(node, {
                 line: operatorToken.loc.end.line,
