+++ lib/rules/space-before-blocks.js
@@ -6,32 +6,14 @@
 
 "use strict";
 
+var astUtils = require("../ast-utils");
+
 //------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
 
 module.exports = function(context) {
     var requireSpace = context.options[0] !== "never";
-
-    /**
-     * Determines whether two adjacent tokens are have whitespace between them.
-     * @param {Object} left - The left token object.
-     * @param {Object} right - The right token object.
-     * @returns {boolean} Whether or not there is space between the tokens.
-     */
-    function isSpaced(left, right) {
-        return left.range[1] < right.range[0];
-    }
-
-    /**
-     * Determines whether two adjacent tokens are on the same line.
-     * @param {Object} left - The left token object.
-     * @param {Object} right - The right token object.
-     * @returns {boolean} Whether or not the tokens are on the same line.
-     */
-    function isSameLine(left, right) {
-        return left.loc.start.line === right.loc.start.line;
-    }
 
     /**
      * Checks the given BlockStatement node has a preceding space if it doesn’t start on a new line.
@@ -42,8 +24,8 @@ module.exports = function(context) {
         var precedingToken = context.getTokenBefore(node),
             hasSpace;
 
-        if (precedingToken && isSameLine(precedingToken, node)) {
-            hasSpace = isSpaced(precedingToken, node);
+        if (precedingToken && astUtils.isTokenOnSameLine(precedingToken, node)) {
+            hasSpace = astUtils.isTokenSpaced(precedingToken, node);
 
             if (requireSpace) {
                 if (!hasSpace) {
